Add files via upload
When the synchronization is performed, a keyboard mark is required for the beginning of the EEG segments. Files named end with number 0-5 are different procedures to generate a summary table of all the treatment designed, participant #, and EEG measurement. Integrated files named Cognitive load_Even/Odd are the combinations of the number files to avoid running several m files for one participant. CSV files are sample data to try, participant # and the name of the eye-tracking segments need to be changed for the corresponding m file you need run. All the eye-tracking files are named p#-learning.csv, and EEG files are named as p#.csv. 

######  EEG files are too large to be uploaded, email me asking for p7.csv  #########